move styling & utility functions on separate files
